adding permissons

+++ api/pitch/delete-pitch.ts
@@ -3,11 +3,15 @@ import Pitch from "../../db/pitch";
 import asyncHandler from "../middlewares/async-handler";
 
 const deletePitch = asyncHandler(async (req, res) => {
-    const { id } = req.params;
+    if (req.user?.role == "Admin") {
+        const { id } = req.params;
 
-    await Pitch.findByIdAndDelete(id)
+        await Pitch.findByIdAndDelete(id)
 
-    res.json({ message: "Pitch deleted Successfully", deleted: true })
+        res.json({ message: "Pitch deleted Successfully", deleted: true })
+    } else {
+        res.status(401).json({ error: "Unauthorized" })
+    }
 })
 
 export default deletePitch;
